設定の型定義を移動
